requirements.txt
filterpy==1.4.5
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
h5py==3.1.0
matplotlib==3.3.4
numba==0.53.1
numpy==1.19.5
nuscenes-devkit==1.1.9
PyYAML==5.4.1
Shapely==1.8.0
tqdm==4.62.3
# waymo-open-dataset-tf-2-1-0==1.3.1
pandas==1.1.5
